refactor: move ApiRequest a top-level module
