+++ src/PostgREST/ApiRequest/QueryParams.hs
@@ -1,12 +1,12 @@
 -- |
--- Module      : PostgREST.Request.QueryParams
+-- Module      : PostgREST.ApiRequest.QueryParams
 -- Description : Parser for PostgREST Query parameters
 --
 -- This module is in charge of parsing all the querystring values in an url, e.g.
 -- the select, id, order in `/projects?select=id,name&id=eq.1&order=id,name.desc`.
 {-# LANGUAGE LambdaCase    #-}
 {-# LANGUAGE TupleSections #-}
-module PostgREST.Request.QueryParams
+module PostgREST.ApiRequest.QueryParams
   ( parse
   , QueryParams(..)
   , pRequestRange
@@ -44,16 +44,18 @@ import PostgREST.RangeQuery              (NonnegRange, allRange,
                                           rangeGeq, rangeLimit,
                                           rangeOffset, restrictRange)
 
-import PostgREST.Request.Types (EmbedParam (..), EmbedPath, Field,
-                                Filter (..), FtsOperator (..),
-                                JoinType (..), JsonOperand (..),
-                                JsonOperation (..), JsonPath, ListVal,
-                                LogicOperator (..), LogicTree (..),
-                                OpExpr (..), Operation (..),
-                                OrderDirection (..), OrderNulls (..),
-                                OrderTerm (..), QPError (..),
-                                SelectItem, SimpleOperator (..),
-                                SingleVal, TrileanVal (..))
+import PostgREST.ApiRequest.Types (EmbedParam (..), EmbedPath, Field,
+                                   Filter (..), FtsOperator (..),
+                                   JoinType (..), JsonOperand (..),
+                                   JsonOperation (..), JsonPath,
+                                   ListVal, LogicOperator (..),
+                                   LogicTree (..), OpExpr (..),
+                                   Operation (..),
+                                   OrderDirection (..),
+                                   OrderNulls (..), OrderTerm (..),
+                                   QPError (..), SelectItem,
+                                   SimpleOperator (..), SingleVal,
+                                   TrileanVal (..))
 
 import Protolude hiding (try)
