+++ src/PostgREST/Query/SqlFragment.hs
@@ -56,13 +56,7 @@ import Control.Arrow ((***))
 import Data.Foldable                 (foldr1)
 import Text.InterpolatedString.Perl6 (qc)
 
-import PostgREST.DbStructure.Identifiers (FieldName,
-                                          QualifiedIdentifier (..))
-import PostgREST.MediaType               (MTPlanFormat (..),
-                                          MTPlanOption (..))
-import PostgREST.RangeQuery              (NonnegRange, allRange,
-                                          rangeLimit, rangeOffset)
-import PostgREST.Request.Types           (Alias, Field, Filter (..),
+import PostgREST.ApiRequest.Types        (Alias, Field, Filter (..),
                                           FtsOperator (..),
                                           JoinCondition (..),
                                           JsonOperand (..),
@@ -76,6 +70,12 @@ import PostgREST.Request.Types           (Alias, Field, Filter (..),
                                           OrderTerm (..), SelectItem,
                                           SimpleOperator (..),
                                           TrileanVal (..))
+import PostgREST.DbStructure.Identifiers (FieldName,
+                                          QualifiedIdentifier (..))
+import PostgREST.MediaType               (MTPlanFormat (..),
+                                          MTPlanOption (..))
+import PostgREST.RangeQuery              (NonnegRange, allRange,
+                                          rangeLimit, rangeOffset)
 
 import Protolude hiding (cast)
